# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Whim 单组织 RBAC API",
    "version": "1.0.0",
    "description": "用户全局唯一，单主部门，多角色、多岗位。数据范围：1全部、2自定义部门、3本部门、4本部门及下级、5本人。岗位不会隐式授予角色。Long 超出 JavaScript 安全整数时序列化为字符串。"
  },
  "servers": [
    {
      "url": "http://localhost:8089"
    }
  ],
...
  "paths": {
...
    "/system/post": {
      "get": {
        "tags": [
          "岗位"
        ],
        "summary": "分页查询本次列表权限可见的岗位。",
        "operationId": "pagePosts",
        "description": "操作权限：`system:post:list`。按该操作实际获授权的角色合并数据范围，详情、修改和删除均检查目标范围。",
        "parameters": [
          {
            "name": "pageNum",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "description": "页码"
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 10
            },
            "description": "每页数量"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultPageDataVOPostVO"
                }
              }
            }
          },
          "400": {
            "description": "参数校验或业务规则不满足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌失效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "403": {
            "description": "缺少操作权限或目标数据不可访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "409": {
            "description": "唯一键冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "429": {
            "description": "请求过于频繁",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "岗位"
        ],
        "summary": "创建系统岗位。",
        "operationId": "createPost",
        "description": "操作权限：`system:post:create`。按该操作实际获授权的角色合并数据范围，详情、修改和删除均检查目标范围。",
        "parameters": [],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultId"
                }
              }
            }
          },
          "400": {
            "description": "参数校验或业务规则不满足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌失效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "403": {
            "description": "缺少操作权限或目标数据不可访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "409": {
            "description": "唯一键冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "429": {
            "description": "请求过于频繁",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostSaveRequest"
              }
            }
          }
        }
      }
    },
    "/system/post/{postId}": {
      "get": {
        "tags": [
          "岗位"
        ],
        "summary": "查询本次详情权限可见的岗位。",
        "operationId": "getPost",
        "description": "操作权限：`system:post:detail`。按该操作实际获授权的角色合并数据范围，详情、修改和删除均检查目标范围。",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^-?[0-9]+$"
            },
            "description": "目标主键 ID"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultPostVO"
                }
              }
            }
          },
          "400": {
            "description": "参数校验或业务规则不满足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌失效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "403": {
            "description": "缺少操作权限或目标数据不可访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "409": {
            "description": "唯一键冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "429": {
            "description": "请求过于频繁",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "岗位"
        ],
        "summary": "修改本次更新权限可操作的岗位。",
        "operationId": "updatePost",
        "description": "操作权限：`system:post:update`。按该操作实际获授权的角色合并数据范围，详情、修改和删除均检查目标范围。",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^-?[0-9]+$"
            },
            "description": "目标主键 ID"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          },
          "400": {
            "description": "参数校验或业务规则不满足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌失效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "403": {
            "description": "缺少操作权限或目标数据不可访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "409": {
            "description": "唯一键冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "429": {
            "description": "请求过于频繁",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostSaveRequest"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "岗位"
        ],
        "summary": "软删本次删除权限可操作的岗位。",
        "operationId": "deletePost",
        "description": "操作权限：`system:post:delete`。按该操作实际获授权的角色合并数据范围，详情、修改和删除均检查目标范围。",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^-?[0-9]+$"
            },
            "description": "目标主键 ID"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          },
          "400": {
            "description": "参数校验或业务规则不满足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌失效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "403": {
            "description": "缺少操作权限或目标数据不可访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "409": {
            "description": "唯一键冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "429": {
            "description": "请求过于频繁",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          }
        }
      }
    },
    "/system/post/{postId}/members": {
      "get": {
        "tags": [
          "岗位"
        ],
        "summary": "查询本次成员查看权限可见岗位的成员。",
        "operationId": "getPostMemberIds",
        "description": "操作权限：`system:post:member:list`。按该操作实际获授权的角色合并数据范围，详情、修改和删除均检查目标范围。",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^-?[0-9]+$"
            },
            "description": "目标主键 ID"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultIdList"
                }
              }
            }
          },
          "400": {
            "description": "参数校验或业务规则不满足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌失效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "403": {
            "description": "缺少操作权限或目标数据不可访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "409": {
            "description": "唯一键冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "429": {
            "description": "请求过于频繁",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "岗位"
        ],
        "summary": "覆盖本次成员分配权限可操作岗位的成员。",
        "operationId": "replacePostMembers",
        "description": "操作权限：`system:post:member:assign`。按该操作实际获授权的角色合并数据范围，详情、修改和删除均检查目标范围。",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^-?[0-9]+$"
            },
            "description": "目标主键 ID"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          },
          "400": {
            "description": "参数校验或业务规则不满足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌失效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "403": {
            "description": "缺少操作权限或目标数据不可访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "409": {
            "description": "唯一键冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "429": {
            "description": "请求过于频繁",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostMemberAssignRequest"
              }
            }
          }
        }
      }
    },
    "/system/post/{postId}/status": {
      "put": {
        "tags": [
          "岗位"
        ],
        "summary": "启用或停用本次状态权限可操作的岗位。",
        "operationId": "setPostStatus",
        "description": "操作权限：`system:post:status`。按该操作实际获授权的角色合并数据范围，详情、修改和删除均检查目标范围。",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^-?[0-9]+$"
            },
            "description": "目标主键 ID"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          },
          "400": {
            "description": "参数校验或业务规则不满足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌失效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "403": {
            "description": "缺少操作权限或目标数据不可访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "409": {
            "description": "唯一键冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "429": {
            "description": "请求过于频繁",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusRequest"
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "PostData": {
        "type": "object",
        "properties": {
          "id": {
            "oneOf": [
              {
                "type": "integer",
                "format": "int64"
              },
              {
                "type": "string",
                "pattern": "^-?[0-9]+$"
              }
            ],
            "description": "Java Long；响应超出 JavaScript 安全整数范围时为字符串，请保持 ID 原始值。"
          },
          "deptId": {
            "oneOf": [
              {
                "type": "integer",
                "format": "int64"
              },
              {
                "type": "string",
                "pattern": "^-?[0-9]+$"
              }
            ],
            "description": "Java Long；响应超出 JavaScript 安全整数范围时为字符串，请保持 ID 原始值。"
          },
          "ownerUserId": {
            "oneOf": [
              {
                "type": "integer",
                "format": "int64"
              },
              {
                "type": "string",
                "pattern": "^-?[0-9]+$"
              }
            ],
            "description": "Java Long；响应超出 JavaScript 安全整数范围时为字符串，请保持 ID 原始值。"
          },
          "postName": {
            "type": "string"
          },
          "postCode": {
            "type": "string"
          },
          "sort": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "remark": {
            "type": "string"
          },
          "createBy": {
            "oneOf": [
              {
                "type": "integer",
                "format": "int64"
              },
              {
                "type": "string",
                "pattern": "^-?[0-9]+$"
              }
            ],
            "description": "Java Long；响应超出 JavaScript 安全整数范围时为字符串，请保持 ID 原始值。"
          },
          "createTime": {
            "type": "string",
            "pattern": "^\\d{4}-\\d{2}-\\d{2} \\d{2}:\\d{2}:\\d{2}$",
            "example": "2026-09-30 12:00:00"
          }
        },
        "description": "系统岗位响应。"
      },
...
      "PostSaveRequest": {
        "type": "object",
        "properties": {
          "postName": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1
          },
          "postCode": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1
          },
          "deptId": {
            "oneOf": [
              {
                "type": "integer",
                "format": "int64"
              },
              {
                "type": "string",
                "pattern": "^-?[0-9]+$"
              }
            ],
            "description": "Java Long；响应超出 JavaScript 安全整数范围时为字符串，请保持 ID 原始值。"
          },
          "sort": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "remark": {
            "type": "string",
            "maxLength": 255
          }
        },
        "required": [
          "postName",
          "postCode",
          "sort"
        ],
        "description": "系统岗位保存参数。"
      },
      "PostMemberAssignRequest": {
        "type": "object",
        "properties": {
          "userIds": {
            "type": "array",
            "items": {
              "oneOf": [
                {
                  "type": "integer",
                  "format": "int64"
                },
                {
                  "type": "string",
                  "pattern": "^-?[0-9]+$"
                }
              ],
              "description": "Java Long；响应超出 JavaScript 安全整数范围时为字符串，请保持 ID 原始值。"
            },
            "uniqueItems": true
          }
        },
        "required": [
          "userIds"
        ],
        "description": "岗位成员覆盖式分配参数。"
      },
...
      "StatusRequest": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "format": "int32"
          }
        },
        "required": [
          "status"
        ],
        "description": "全局权限状态修改参数。"
      },
      "ErrorResult": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "与 HTTP 状态码一致。"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "nullable": true,
            "oneOf": [
              {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "field": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            ]
          },
          "success": {
            "type": "boolean",
            "example": false
          }
        },
        "required": [
          "code",
          "message",
          "success"
        ]
      },
...
      "ResultVoid": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string"
          },
          "data": {
            "nullable": true,
            "description": "无返回数据，值为 null。"
          },
          "success": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "code",
          "message",
          "data",
          "success"
        ]
      },
...
      "ResultId": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string"
          },
          "data": {
            "oneOf": [
              {
                "type": "integer",
                "format": "int64"
              },
              {
                "type": "string",
                "pattern": "^-?[0-9]+$"
              }
            ],
            "description": "Java Long；响应超出 JavaScript 安全整数范围时为字符串，请保持 ID 原始值。"
          },
          "success": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "code",
          "message",
          "data",
          "success"
        ]
      },
...
      "ResultIdList": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "oneOf": [
                {
                  "type": "integer",
                  "format": "int64"
                },
                {
                  "type": "string",
                  "pattern": "^-?[0-9]+$"
                }
              ],
              "description": "Java Long；响应超出 JavaScript 安全整数范围时为字符串，请保持 ID 原始值。"
            },
            "uniqueItems": true
          },
          "success": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "code",
          "message",
          "data",
          "success"
        ]
      },
...
      "PagePostData": {
        "type": "object",
        "properties": {
          "currentPage": {
            "oneOf": [
              {
                "type": "integer",
                "format": "int64"
              },
              {
                "type": "string",
                "pattern": "^-?[0-9]+$"
              }
            ],
            "description": "Java Long；响应超出 JavaScript 安全整数范围时为字符串，请保持 ID 原始值。"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PostData"
            }
          },
          "pages": {
            "oneOf": [
              {
                "type": "integer",
                "format": "int64"
              },
              {
                "type": "string",
                "pattern": "^-?[0-9]+$"
              }
            ],
            "description": "Java Long；响应超出 JavaScript 安全整数范围时为字符串，请保持 ID 原始值。"
          },
          "size": {
            "oneOf": [
              {
                "type": "integer",
                "format": "int64"
              },
              {
                "type": "string",
                "pattern": "^-?[0-9]+$"
              }
            ],
            "description": "Java Long；响应超出 JavaScript 安全整数范围时为字符串，请保持 ID 原始值。"
          },
          "total": {
            "oneOf": [
              {
                "type": "integer",
                "format": "int64"
              },
              {
                "type": "string",
                "pattern": "^-?[0-9]+$"
              }
            ],
            "description": "Java Long；响应超出 JavaScript 安全整数范围时为字符串，请保持 ID 原始值。"
          }
        },
        "required": [
          "currentPage",
          "data",
          "pages",
          "size",
          "total"
        ]
      },
      "ResultPageDataVOPostVO": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/PagePostData"
          },
          "success": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "code",
          "message",
          "data",
          "success"
        ]
      },
      "ResultPostVO": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/PostData"
          },
          "success": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "code",
          "message",
          "data",
          "success"
        ]
      }
    }
  }
}